docs(example): Split s-expression example
